OJ-1825: test folder restructure
