+++ integration-tests/tests/aws/resources/nino-check-dynamodb-helper.ts
