+++ integration-tests/tests/aws/resources/nino-check-stepfunction-helper.ts
